% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxmix-package.R
\docType{package}
\name{boxcoxmix-package}
\alias{boxcoxmix-package}
\alias{boxcoxmix}
\title{Box-Cox-Type Transformations for Linear and Logistic Models with Random Effects}
\description{
Box-Cox-type transformations for linear and logistic models with
random effects using non-parametric profile maximum likelihood
estimation. The main functions are \code{optim.boxcox()} for linear models with
random effects and \code{boxcoxtype()} for logistic models with
random effects.
}
\details{
\tabular{ll}{ Package: \tab boxcoxmix\cr Type: \tab Package\cr Version: \tab
0.28\cr Date: \tab 2020-09-17\cr License: \tab GPL (>=3)\cr }
}
\references{
Box G. and Cox D. (1964). An analysis of transformations. Journal of
the Royal Statistical Society. Series B (Methodological), pages 211-252.

Aitkin, M. A., Francis, B., Hinde, J., and Darnell, R. (2009). Statistical
modelling in R. Oxford University Press Oxford.

Jochen Einbeck, Ross Darnell and John Hinde (2014). npmlreg:
Nonparametric maximum likelihood estimation for random effect
models. R package version 0.46-1.

R Core Team (2016). R: A language and environment for
statistical computing. R Foundation for Statistical Computing,
Vienna, Austria. URL https://www.R-project.org/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iagogv/boxcoxmix}
  \item Report bugs at \url{https://gitlab.com/iagogv/boxcoxmix/-/issues}
}

}
\author{
\strong{Maintainer}: Iago Giné-Vázquez \email{iago.gin-vaz@protonmail.com} (\href{https://orcid.org/0000-0002-6725-2638}{ORCID})

Authors:
\itemize{
  \item Amani Almohaimeed \email{ama.almohaimeed@qu.edu.sa} (\href{https://orcid.org/0000-0002-3205-0921}{ORCID})
  \item Jochen Einbeck \email{jochen.einbeck@durham.ac.uk} (\href{https://orcid.org/0000-0002-9457-2020}{ORCID})
}

}
\keyword{package}
