% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfind.R
\name{Kfind.boxcox}
\alias{Kfind.boxcox}
\title{Grid search over K for NPML estimation of random effect and variance component models}
\usage{
Kfind.boxcox(
  formula,
  groups = 1,
  data,
  lambda = 1,
  EMdev.change = 1e-04,
  steps = 500,
  find.k = c(2, 10),
  model.selection = "aic",
  start = "gq",
  find.tol = c(0, 1.5),
  steps.tol = 15,
  ...
)
}
\arguments{
\item{formula}{a formula describing the transformed response and the fixed
effect model (e.g. y ~ x).}

\item{groups}{the random effects. To fit overdispersion models , set \code{groups} = 1.}

\item{data}{a data frame containing variables used in the fixed and random
effect models.}

\item{lambda}{a transformation parameter, setting \code{lambda}=1 means 'no
transformation'.}

\item{EMdev.change}{a small scalar, with default 0.0001, used to determine
when to stop EM algorithm.}

\item{steps}{maximum number of iterations for the EM algorithm.}

\item{find.k}{search in a range of \code{K}, with default (2,10)
in step of 1.}

\item{model.selection}{Set \code{model.selection="aic"}, to use Akaike information criterion 
as model selection criterion or \code{model.selection="bic"}, to use Bayesian information criterion 
as model selection criterion.}

\item{start}{a description of the initial values to be used in the fitted
model, Quantile-based version "quantile" or Gaussian Quadrature "gq" can be
set.}

\item{find.tol}{search in a range of \code{tol}, with default (0,1.5)
in step of 1.}

\item{steps.tol}{number of points in the grid search of \code{tol}.}

\item{\dots}{extra arguments will be ignored.}
}
\value{
List with class \code{boxcoxmix} containing:
\item{MinDisparity}{the minimum disparity found.} \item{Best.K}{the
value of \code{K} corresponding to \code{MinDisparity}.}
\item{AllMinDisparities }{a vector containing all minimum disparities calculated on the
grid.} \item{AllMintol }{list of \code{tol} values used in the grid.}
\item{All.K }{list of \code{K} values used in the grid.}
\item{All.aic}{the Akaike information criterion of all fitted regression models.}
 \item{All.bic}{the Bayesian information criterion of all fitted regression models.}
}
\description{
A grid search over the parameter \code{K}, to set the best number of
mass-points.
}
\details{
Not only the shape of the distribution causes the skewness it may due to the use of an 
insufficient number of classes, \code{K}. For this, the \code{Kfind.boxcox()} function 
was created to search over a selected range of \code{K} and find the best. For each number
 of classes, a grid search over \code{tol} is performed and the \code{tol} with the lowest 
 \code{aic} or \code{bic} value is considered as the optimal. Having the minimal \code{aic} or \code{bic} values for a whole range of
  \code{K} that have been specified beforehand, the \code{Kfind.boxcox()} function can find 
  the best number of the component as the one with the smallest value. It also plots the \code{aic} or \code{bic} values against
 the selected range of \code{K}, including a vertical line indicating the best value of \code{K}
 that minimizes the model selection criteria. The full range of 
  classes and their corresponding optimal \code{tol} can be printed off from the \code{Kfind.boxcox()}'s
   output and used with other \pkg{boxcoxmix} functions as starting points.
}
\examples{
# Fabric data
\donttest{data(fabric, package = "npmlreg")
teststr<-Kfind.boxcox(y ~ x, data = fabric,  start = "gq", groups=1, 
find.k = c(2, 3), model.selection = "aic", steps.tol=5)
# Minimal AIC: 202.2114 at K= 2 }









}
\seealso{
\code{\link{tolfind.boxcox}}.
}
\author{
Amani Almohaimeed and Jochen Einbeck
}
\keyword{Kfind}
\keyword{boxcox}
