% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.boxcoxmix.R
\name{print.boxcoxmix}
\alias{print.boxcoxmix}
\alias{summary.boxcoxmix}
\alias{print.boxcoxmixpure}
\alias{summary.boxcoxmixpure}
\title{Summary of boxcoxmix functions}
\usage{
\method{print}{boxcoxmix}(x, digits = max(3, getOption("digits") - 3), na.print = "", ...)

\method{print}{boxcoxmixpure}(x, digits = max(3, getOption("digits") - 3), na.print = "", ...)

\method{summary}{boxcoxmix}(object, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{boxcoxmixpure}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object for which a summary is desired.}

\item{digits}{an integer number format.}

\item{na.print}{a character string which is used to indicate NA values output format.}

\item{\dots}{additional arguments.}

\item{object}{an object for which a summary is desired.}
}
\value{
Return invisibly the main object provided, while print a summary of its content.
}
\description{
\code{summary()} and \code{print()} are generic functions used to produce the results of the functions:
\code{np.boxcoxmix()}, \code{optim.boxcox()} and \code{tolfind.boxcox()}.
}
