% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxmix-package.R
\docType{package}
\name{boxcoxmix-package}
\alias{boxcoxmix-package}
\alias{boxcoxmix}
\title{Response Transformations for Random Effect and Variance Component Models}
\description{
Response transformations for overdispersed generalized linear models and
variance component models using nonparametric profile maximum likelihood
estimation. The main function is \code{optim.boxcox()}.
}
\details{
\tabular{ll}{ Package: \tab boxcoxmix\cr Type: \tab Package\cr Version: \tab
0.14\cr Date: \tab 2017-10-9\cr License: \tab GPL (>=3)\cr }
}
\references{
Box G. and Cox D. (1964). An analysis of transformations. Journal of
the Royal Statistical Society. Series B (Methodological), pages 211-252.

Aitkin, M. A., Francis, B., Hinde, J., and Darnell, R. (2009). Statistical
modelling in R. Oxford University Press Oxford.

Jochen Einbeck, Ross Darnell and John Hinde (2014). npmlreg:
Nonparametric maximum likelihood estimation for random effect
models. R package version 0.46-1.

Maintainer: Amani Almohaimeed <amani.almohaimeed@gmail.com>
}
\author{
Amani Almohaimeed and Jochen Einbeck
}
\keyword{package}
